Previous and next links shouldn't have an active class
It's not logically to give your previous and next links an active class. You can't 'be' on a previous or next link so therefor it can't be active. Previous or next links are only visual indicators for moving to the next or previous page.

This was also conflicting with Twitter Bootstrap as it showed the link as non-clickable while it should just display as a normal clickable link.

Signed-off-by: Dries Vints <dries.vints@gmail.com>